Factor BCP-004-01 tests into dedicated test suite.
